Functions have been divided into modules
