+++ js/util.js
@@ -1,0 +1,28 @@
+const getRandomInt = (start, end) => {
+  if (start >= end) {
+    throw new Error('The "start" value must be less than the "end" value.');
+  } else if (start < 0 || end < 0) {
+    throw new Error('A range with negative numbers is not allowed.');
+  } else {
+    const min = Math.ceil(start);
+    const max = Math.floor(end);
+    return Math.floor(Math.random() * (max - min + 1)) + min;
+  }
+};
+
+const checkStringLength = (testLine, maxLength) => testLine.length < maxLength;
+
+const getRandomObject = (objects) => objects[getRandomInt(0, objects.length - 1)];
+
+const getId = (array) => {
+  for (let i = 0; i < array.length; i++) {
+    if (array[i]) {
+      array[i] = false;
+      return i + 1;
+    }
+  }
+};
+
+checkStringLength('Goodbye world!', 15);
+
+export {getRandomInt, getRandomObject, getId};
